Add item_size, width and height to LocationSection
Also moves remaining LocationSection stuff out of location.cpp
and minor cleanup of touched code.
