+++ src/core/location.cpp
@@ -2,10 +2,13 @@
 #include <lua.h>
 #include <luaglue/luainterface.h>
 #include "jsonutil.h"
+#include "rule.h"
 #include "tracker.h"
 #include "util.h"
 
+
 using nlohmann::json;
+
 
 const LuaInterface<Location>::MethodMap Location::Lua_Methods = {};
 
@@ -29,33 +32,6 @@ int Location::Lua_Index(lua_State *L, const char *key)
 bool Location::Lua_NewIndex(lua_State *L, const char *key)
 {
     return false;
-}
-
-
-static bool parseRule(const json& v, std::list<std::string>& rule,
-        const char* nodeType, const char* ruleType, const std::string& name)
-{
-    if (v.is_string()) {
-        // string with individual codes separated by comma
-        commasplit(v, rule);
-    }
-    else if (v.is_array()) {
-        // we also allow rules to be arrays of strings instead
-        for (const auto& part: v) {
-            if (!part.is_string()) {
-                fprintf(stderr, "%s: bad %s rule in \"%s\"\n",
-                    nodeType, ruleType, sanitize_print(name).c_str());
-                continue;
-            }
-            rule.push_back(part);
-        }
-    }
-    else {
-        fprintf(stderr, "%s: bad %s rule in \"%s\"\n",
-            nodeType, ruleType, sanitize_print(name).c_str());
-        return false;;
-    }
-    return true;
 }
 
 
@@ -225,6 +201,7 @@ std::list<Location> Location::FromJSON(json& j, const std::list<Location>& paren
     return locs;
 }
 
+
 void Location::merge(const Location& other)
 {
     for (auto& maploc: other._mapLocations) {
@@ -235,7 +212,6 @@ void Location::merge(const Location& other)
         _sections.push_back(sec);
     }
 }
-
 
 Location::MapLocation Location::MapLocation::FromJSON(json& j)
 {
@@ -281,162 +257,6 @@ Location::MapLocation Location::MapLocation::FromJSON(json& j)
 
     return maploc;
 }
-
-
-LocationSection LocationSection::FromJSON(json& j, const std::string& parentId, const std::list< std::list<std::string> >& parentAccessRules, const std::list< std::list<std::string> >& parentVisibilityRules, const std::string& closedImg, const std::string& openedImg, const std::string& overlayBackground)
-{
-    // TODO: pass inherited values as parent instead
-    LocationSection sec;
-    sec._parentId = parentId;
-    sec._name = to_string(j["name"],sec._name);
-    sec._clearAsGroup = to_bool(j["clear_as_group"],sec._clearAsGroup);
-    sec._closedImg = to_string(j["chest_unopened_img"], closedImg);
-    sec._openedImg = to_string(j["chest_opened_img"], openedImg);
-    sec._overlayBackground = to_string(j["overlay_background"], overlayBackground);
-    auto tmp = to_string(j["hosted_item"], "");
-    commasplit(tmp, sec._hostedItems);
-    sec._ref = to_string(j, "ref", "");
-    sec._itemCount = sec._hostedItems.empty() && sec._ref.empty() ? 1 : 0;
-    sec._itemCount = to_int(j["item_count"], sec._itemCount);
-    bool nonEmpty = sec._itemCount > 0 || !sec._hostedItems.empty();
-
-    if (j["access_rules"].is_array() && !j["access_rules"].empty()) {
-        // TODO: merge code with Location's access rules
-        nonEmpty = true;
-        for (const auto& v : j["access_rules"]) {
-            std::list<std::string> newRule;
-            if (!parseRule(v, newRule, "LocationSection", "access", sec._name))
-                continue;
-            for (auto oldRule : parentAccessRules) {
-                for (auto& newTest : newRule) {
-                    oldRule.push_back(newTest);
-                }
-                sec._accessRules.push_back(oldRule);
-            }
-            if (parentAccessRules.empty()) {
-                sec._accessRules.push_back(newRule);
-            }
-        }
-    } else if (j["access_rules"].is_string() && !j["access_rules"].empty()) {
-        // single string access rule, same as [["code"]]
-        const std::string& newTest = j["access_rules"];
-        for (auto oldRule : parentAccessRules) {
-            oldRule.push_back(newTest);
-            sec._accessRules.push_back(oldRule);
-        }
-        if (parentAccessRules.empty()) {
-            sec._accessRules.push_back({newTest});
-        }
-    } else {
-        sec._accessRules = parentAccessRules;
-        if (!j["access_rules"].is_null() && !j["access_rules"].is_array()) {
-            fprintf(stderr, "Location: Section: invalid access rules in \"%s\"\n",
-                    sanitize_print(sec._name).c_str());
-        }
-    }
-    if (j["visibility_rules"].is_array() && !j["visibility_rules"].empty()) {
-        // TODO: merge code with Location's access rules
-        nonEmpty = true;
-        for (const auto& v : j["visibility_rules"]) {
-            std::list<std::string> newRule;
-            if (!parseRule(v, newRule, "LocationSection", "visibility", sec._name))
-                continue;
-            for (auto oldRule : parentVisibilityRules) {
-                for (auto& newTest : newRule) {
-                    oldRule.push_back(newTest);
-                }
-                sec._visibilityRules.push_back(oldRule);
-            }
-            if (parentVisibilityRules.empty()) {
-                sec._visibilityRules.push_back(newRule);
-            }
-        }
-    } else if (j["visibility_rules"].is_string() && !j["visibility_rules"].empty()) {
-        // single string visibility rule, same as [["code"]]
-        const std::string& newTest = j["visibility_rules"];
-        for (auto oldRule : parentVisibilityRules) {
-            oldRule.push_back(newTest);
-            sec._visibilityRules.push_back(oldRule);
-        }
-        if (parentVisibilityRules.empty()) {
-            sec._visibilityRules.push_back({newTest});
-        }
-    } else {
-        sec._visibilityRules = parentVisibilityRules;
-        if (!j["visibility_rules"].is_null() && !j["visibility_rules"].is_array()) {
-            fprintf(stderr, "Location: Section: invalid visibility rules in \"%s\"\n",
-                    sanitize_print(sec._name).c_str());
-        }
-    }
-
-    if (!sec._ref.empty() && nonEmpty) {
-        fprintf(stderr, "Location: Section: extra data in section \"%s\" with \"ref\"\n",
-                sanitize_print(sec._name).c_str());
-    }
-
-    return sec;
-}
-
-
-bool LocationSection::clearItem(bool all)
-{
-    if (_itemCleared >= _itemCount) return false;
-    if (_clearAsGroup || all)
-        _itemCleared = _itemCount;
-    else
-        _itemCleared++;
-    onChange.emit(this);
-    return true;
-}
-bool LocationSection::unclearItem()
-{
-    if (_itemCleared == 0) return false;
-    if (_clearAsGroup)
-        _itemCleared = 0;
-    else
-        _itemCleared--;
-    onChange.emit(this);
-    return true;
-}
-
-json LocationSection::save() const
-{
-    json j = json::object();
-    if (_itemCleared)
-        j["cleared"] = _itemCleared;
-    if (_highlight != Highlight::NONE)
-        j["highlight"] = HighlightToString(_highlight);
-    return j;
-}
-
-bool LocationSection::load(json& j)
-{
-    if (j.type() == json::value_t::object) {
-        bool changed = false;
-        int val = to_int(j["cleared"], _itemCleared);
-        if (val != _itemCleared) {
-            _itemCleared = val;
-            changed = true;
-        }
-        const auto highlight = HighlightFromString(to_string(j, "highlight", ""));
-        if (highlight != _highlight) {
-            _highlight = highlight;
-            changed = true;
-        }
-        if (changed)
-            onChange.emit(this);
-        return true;
-    }
-    return false;
-}
-
-bool LocationSection::operator<(const LocationSection& rhs) const
-{
-    if (this->getParentID() == rhs.getParentID())
-        return this->getName() < rhs.getName();
-    return this->getParentID() < rhs.getParentID();
-}
-
 
 #ifndef NDEBUG
 #include <stdio.h>
